% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\docType{class}
\name{Backend.rds-class}
\alias{Backend.rds}
\alias{Backend.rds-class}
\title{rds backend class}
\description{
rds backend class
}
\section{Fields}{

\describe{
\item{\code{dir}}{storage directory}
}}

