% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend.rds}
\alias{backend.rds}
\title{create rds backend}
\usage{
backend.rds(dir = "./data")
}
\arguments{
\item{...}{arguments passed to Backend.rds$new()}
}
\value{
a rds backend
}
\description{
create rds backend
}

