% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{create.backend}
\alias{create.backend}
\title{create backend}
\usage{
create.backend(name = "rds", dir = "./data")
}
\arguments{
\item{...}{arguments passed to backend constructor}
}
\value{
a rds backend
}
\description{
create backend
}

