% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{forget.obj}
\alias{forget.obj}
\title{Completely remove the object from persistent storage including all revisions.}
\usage{
forget.obj(name, backend = NULL)
}
\arguments{
\item{name}{object name on the storage.}

\item{backend}{backend storage.}
}
\value{
the name of forgot objects.
}
\description{
Completely remove the object from persistent storage including all revisions.
}

