% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{forget.rev}
\alias{forget.rev}
\title{remove all object of specific revision from persistent storage.}
\usage{
forget.rev(rev, backend = NULL)
}
\arguments{
\item{rev}{revision name of the object.}

\item{backend}{backend storage.}
}
\value{
the name of removed object.
}
\description{
remove all object of specific revision from persistent storage.
}

