% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get.latest.rev}
\alias{get.latest.rev}
\title{create revision name from config object.}
\usage{
get.latest.rev(name, backend = NULL)
}
\arguments{
\item{name}{object name on the storage.}

\item{backend}{backend storage.}
}
\value{
latest revision name.
}
\description{
create revision name from config object.
}

