% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{list.obj}
\alias{list.obj}
\title{list object names and revisions.}
\usage{
list.obj(name = NULL, rev = NULL, backend = NULL)
}
\arguments{
\item{name}{object name on the storage. this argment can be NULL and then the function returns all objects.}

\item{rev}{revision name of the object. this argment can be NULL and then the function returns all revisions.}

\item{backend}{backend storage.}
}
\value{
the name of removed object.
}
\description{
list object names and revisions.
}

