% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{load.obj}
\alias{load.obj}
\title{Load a object from persistent storage.}
\usage{
load.obj(name, rev = NULL, backend = NULL)
}
\arguments{
\item{name}{object name on the storage.}

\item{rev}{revision name of the object. this argment can be NULL and then the function loads the object of the latest revision}

\item{backend}{backend storage.}
}
\value{
loaded object.
}
\description{
Load a object from persistent storage.
}

