% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{load.rev.info}
\alias{load.rev.info}
\title{load revision info.}
\usage{
load.rev.info(rev, backend = NULL)
}
\arguments{
\item{rev}{revision name.}

\item{backend}{backend storage.}
}
\value{
loaded revision info.
}
\description{
load revision info.
}

