% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{obj.exists}
\alias{obj.exists}
\title{check if object exists.}
\usage{
obj.exists(name, rev = NULL, backend = NULL)
}
\arguments{
\item{name}{object name.}

\item{rev}{revision name.}

\item{backend}{backend storage.}
}
\value{
TRUE if object exists else FALSE.
}
\description{
check if object exists.
}

