% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{remove.obj}
\alias{remove.obj}
\title{remove the object of specific revesion from persistent storage.}
\usage{
remove.obj(name, rev, backend = NULL)
}
\arguments{
\item{name}{object name on the storage.}

\item{rev}{revision name of the object. this argment can be NULL and then the function loads the object of the latest revision}

\item{backend}{backend storage.}
}
\value{
the name of removed object.
}
\description{
remove the object of specific revesion from persistent storage.
}

