% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{save.obj}
\alias{save.obj}
\title{Save a object to persistent storage.}
\usage{
save.obj(obj, name, rev = NULL, backend = NULL)
}
\arguments{
\item{obj}{object to save.}

\item{name}{object name on the storage.}

\item{rev}{revision name of the object. this argment can be NULL and then revision is determined automatically.}

\item{backend}{backend storage.}
}
\value{
character vector of length 2(name and rev)
}
\description{
Save a object to persistent storage.
}

