% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{save.rev.info}
\alias{save.rev.info}
\title{save revision info.}
\usage{
save.rev.info(config, rev, backend = NULL)
}
\arguments{
\item{backend}{backend storage.}

\item{object}{an object specifies the revision.}
}
\value{
the list of length 2 (name, rev)
}
\description{
save revision info.
}

