% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get.rev.info}
\alias{get.rev.info}
\title{create revision name from config object.}
\usage{
get.rev.info(config = NULL, backend = NULL)
}
\arguments{
\item{config}{config object specifies the revision.}

\item{backend}{backend storage.}
}
\value{
the list of length 2 (rev, revision info)
}
\description{
create revision name from config object.
}

